Use GitHub issue import API to import bugs and comments.
Advantages:
- higher rate limits because creating an issue does not trigger notifications;
- issue, labels, state, and comments imported in a single request;
- comments from the original bugs become comments on the imported issue;
- dates of bug creation and comments are preserved.

Changes initially made by myself for the Bugzilla to GitHub migration of the Coq project
(see https://www.theozimmermann.net/2017/10/bugzilla-to-github/).
Extracted and submitted as a patch on the original script by Martin Michlmayr.
Reduced to a minimally invasive patch by myself to ease integration in the original script.

Co-authored-by: Martin Michlmayr <tbm@cyrius.com>