bugzilla2github: Always use next available GH issue ID
Updates the algorithm to add a new issue to start at ID 1 and simply
increment monotonically looking for the next available ID that does
not exist.

This should simply find the last ID currently active and incrementally
apply issues from there. But, an issue may be created during the
script operation and it will take the next available ID.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>