Track unconverted BZ logins separately
Currently, any BZ login that cannot be converted to a GitHub login
is printed. This is knowingly going to be many logins and the ones
that will be converted will be explicitly specified. Therefore,
clean up script output by not logging them.

The separate set can be logged if desired.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>